[Misc] Refactor models
